refactor(theme): merge duplicate auth logger into forum logger
